[Vimrc] Orgmode - functions implemented
%%% Branch
 |
 ----------> master

%%% Code status
 |
 ----------> Compiling

%%% Level of test
 |
 ----------> Hand-tested

%%% Comments
 |
 ------
      |
      v
The features of vim-orgmode associated with the tags were not fully implemented
yet. The main lack concerned listing and filtering applied to tags.
Some functions, written in vimscript, have been implemented those features. On
top of that, they have been map to <F9> and <F10> in the buffers which filetype
is "org".