[Snippet] Init. of the vim-snippet plugin
%%% Branch
 |
 ----------> vim-snippet

%%% Code status
 |
 ----------> In development
 |
 ----------> Compiling
      |
      v
Even if the code is still in development, the first part which is pushed to the
repository by the present commit is compiling.

%%% Level of test
 |
 ----------> Hand-tested

%%% Comments
 |
 ------
      |
      v
At this point, the installation procedure of the plugin has not been automated.
The word "plugin" is used according to its first meaning. In other words, this
plugin is not (yet) developed to be released (through Pathogen or Vundle).
However, it is implemented as a file in "~/.vim/plugin" since it is loaded at
the initialization of Vim.