[Snippet] Impl. of the plugin [1]
%%% Branch
 |
 ----------> vim-snippet

%%% Code status
 |
 ----------> In development
 |
 ----------> Compiling

%%% Level of test
 |
 ----------> Hand-tested
      |
      v
Hand-tested, but quite thoroughly hand-tested so far.

%%% Comments
 |
 ------
      |
      v
The core of the plugin has been implemented and is functioning.
The installation and integration to Vim has not yet been done.
The features still to be implemented concern the termination of the plugin, and
the potential adding of an endpoint to the snippets.